refactor(team): extract misc HTTP handlers into broker_misc_handlers.go
Slice 7 of broker.go decomposition. Moves handleHealth,
resolveProviderModel, handleSessionMode, handleFocusMode, handleReset,
handleResetDM, respawnAgentPane, handleSignals, handleDecisions,
handleWatchdogs, handleActions, handleQueue, and handleTelegramGroups
into a new broker_misc_handlers.go.

Relocates TestBrokerSessionModePersistsAndSurvivesReset,
TestBrokerActionsAndSchedulerEndpoints, TestBrokerPostsAndDedupesNexNotifications,
and TestQueueEndpointShowsDueJobs into broker_misc_handlers_test.go.
Adds four new characterization tests:
TestHandleHealth_ReturnsOKAndVersionFields,
TestHandleResetDM_RequiresPOST,
TestHandleWatchdogs_GETReturnsCurrentAlerts,
TestHandleQueue_ReturnsDueJobsArray.

handleNexNotifications and the handleMessages stub stay in broker.go
for slice 9 (messages) since they thin-wrap PostAutomationMessage and
handlePostMessage. handleScheduler and handleBridge are out of scope
for this slice.

Total package coverage rises from 63.1% to 63.2%.
